perf(site): draw hero shader before first paint
Replace the CityShader React island with a synchronous inline WebGL boot:
the engine (hero-shader.inline.js) plus the extracted city fragment shader
(city.frag.glsl) are inlined into index.astro and run during HTML parse,
so the backdrop is part of the first paint instead of popping in seconds
later after hydration. Drop the 2.2s in-shader intro ramp and the 300ms
canvas fade; keep the data-ready gate that masks pre-first-frame garbage.
Remove the now-dead React shader components (alternate looks live in git
history).
